+++ example/chaos/model/grid2by2.dart
@@ -1,0 +1,5 @@
+part of chaos;
+
+class Grid2By2 extends Grid {
+  Grid2By2() : super(2, 2);
+}