+++ lib/boarding_model.dart
@@ -4,6 +4,7 @@ import 'dart:math';
 
 part 'model/cell.dart';
 part 'model/grid.dart';
+part 'model/pieces.dart';
 
 part 'util/color.dart';
 part 'util/random.dart';